requirements.txt
streamlit == 1.24.0
streamlit-drawable-canvas
pillow
numpy
